MolVis,VTK: requested changes
