+++ misc/vtk/src/processors/vtkdatasetinformation.cpp
@@ -42,6 +42,7 @@
 #include <warn/pop>
 
 namespace inviwo {
+
 // The Class Identifier has to be globally unique. Use a reverse DNS naming scheme
 const ProcessorInfo VTKDataSetInformation::processorInfo_{
     "org.inviwo.VTKDataSetInformation",  // Class identifier
@@ -51,6 +52,45 @@ const ProcessorInfo VTKDataSetInformation::processorInfo_{
     Tags::CPU,                           // Tags
 };
 const ProcessorInfo VTKDataSetInformation::getProcessorInfo() const { return processorInfo_; }
+
+VTKDataSetInformation::ArrayInformationProperty::ArrayInformationProperty(
+    const std::string& arrayName, const std::string& identifier, vtkDataArray* array)
+    : CompositeProperty(identifier, arrayName)
+    , dataType_(identifier + "dataType", "Data type", array->GetDataTypeAsString())
+    , numberOfComponents_(identifier + "numberOfComponents", "Components",
+                          std::to_string(array->GetNumberOfComponents()))
+    , componentInformation_("componentInformation", "Component info") {
+    addProperties(dataType_, numberOfComponents_, componentInformation_);
+
+    dataType_.setReadOnly(true);
+    numberOfComponents_.setReadOnly(true);
+
+    for (auto i{0}; i < array->GetNumberOfComponents(); ++i) {
+        auto compInfo =
+            new CompositeProperty(fmt::format("component{}", i), fmt::format("Component {}", i));
+
+        const auto componentName = array->HasAComponentName() ? array->GetComponentName(i)
+                                                              : fmt::format("component{}", i + 1);
+        auto name = new StringProperty(fmt::format("name{}", i), "Name", componentName);
+        name->setReadOnly(true);
+        auto numValues = new StringProperty(fmt::format("numVal{}", i), "Number of tuples",
+                                            std::to_string(array->GetNumberOfTuples()));
+        numValues->setReadOnly(true);
+
+        compInfo->addProperty(name);
+        compInfo->addProperty(numValues);
+
+        componentInformation_.addProperty(compInfo);
+    }
+}
+
+const std::string& VTKDataSetInformation::ArrayInformationProperty::getDataType() const {
+    return dataType_.get();
+}
+
+const std::string& VTKDataSetInformation::ArrayInformationProperty::getNumberOfComponents() const {
+    return numberOfComponents_.get();
+}
 
 VTKDataSetInformation::VTKDataSetInformation()
     : Processor()
